Move initializers to its own folder
