+++ client/app/lib/initializers/ace-editor.js
