+++ client/app/lib/initializers/confirm-dialog.jsx
